chore: move hang-breadcrumb contract under diagnostics/
Keeps src/shared/contracts direct file count within allowlist (50).
