+++ src/shared/contracts/diagnostics/hang-breadcrumb.ts
